Dist fixes Also
- Fix crash when the mouse no longer hovers over a StateMachineState node

+++ Saturn/src/Saturn/Animation/Pose.h
@@ -53,7 +53,7 @@ namespace Saturn {
 		float Duration = 0.0f;
 		float Timestamp = 0.0f;
 		// Number of bones
-		uint32_t BonesUsed = 0u;
+		uint64_t BonesUsed = 0llu;
 	};
 	
 }
